test: graduate the #311 repro spec into smoke with coverage tags (#312)
Green with no test change once the fix landed — that's the regression
proof. Moved to e2e/smoke/authenticated/ with @page/@uc tags and a new
mobile-no-horizontal-overflow use-case; keeps the dedicated user and the
chromium viewport override. mode 'default' pins both tests to one worker
since smoke runs fullyParallel and they share a worker-scoped seed.
test:e2e:repro gets --pass-with-no-tests now that the tier is empty;
root CLAUDE.md exemplar pointer follows the file.
